add tests for handlers

+++ ProgressAcademy.Handlers/Commands/LessonCommandHandler.cs
@@ -13,16 +13,29 @@ public class LessonCommandHandler {
 
     public void Handle(CreateLessonCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"CreateLessonCommand must not be null");
+        }
+
         _lessonRepository.CreateLesson(command.Lesson);
     }
 
     public void Handle(UpdateLessonCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"UpdateLessonCommand must not be null");
+        }
         _lessonRepository.UpdateLesson(command.Lesson);
     }
 
     public void Handle(DeleteLessonCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"DeleteLessonCommand must not be null");
+        }
         _lessonRepository.DeleteLesson(command.LessonId);
     }
 }
