+++ ProgressAcademy.Handlers/Commands/MentorCommandHandler.cs
@@ -13,16 +13,31 @@ public class MentorCommandHandler {
 
     public void Handle(CreateMentorCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"CreateMentorCommand must not be null");
+        }
+
         _mentorRepository.CreateMentor(command.Mentor);
     }
 
     public void Handle(UpdateMentorCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"UpdateMentorCommand must not be null");
+        }
+
         _mentorRepository.UpdateMentor(command.Mentor);
     }
 
     public void Handle(DeleteMentorCommand command)
     {
+        if(command == null)
+        {
+            throw new ArgumentNullException($"DeleteMentorCommand must not be null");
+        }
+
         _mentorRepository.DeleteMentor(command.MentorId);
     }
 }
